Fix dialog structure
